Rename TradesHistorian to EventsHistorian
