+++ rotkehlchen/api/server.py
@@ -42,6 +42,7 @@ from rotkehlchen.api.v1.resources import (
     ExchangesResource,
     ExternalServicesResource,
     HistoryDownloadingResource,
+    HistoryStatusResource,
     HistoryExportingResource,
     HistoryProcessingResource,
     IgnoredActionsResource,
@@ -134,6 +135,7 @@ URLS_V1: URLS = [
     ('/messages/', MessagesResource),
     ('/periodic/', PeriodicDataResource),
     ('/history/', HistoryProcessingResource),
+    ('/history/status', HistoryStatusResource),
     ('/history/export/', HistoryExportingResource),
     ('/history/download/', HistoryDownloadingResource),
     ('/queried_addresses', QueriedAddressesResource),
