+++ rotkehlchen/history/__init__.py
@@ -1,2 +1,2 @@
 from .price import PriceHistorian  # noqa: F401
-from .trades import TradesHistorian  # noqa: F401
+from .trades import EventsHistorian  # noqa: F401
